backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langgraph
langchain-groq
sqlalchemy

# For Postgres
psycopg2-binary

# For MySQL (optional)
pymysql
